+++ packages/lucid/src/tx-builder/internal/CompleteTxBuilder.ts
@@ -1,9 +1,16 @@
-import { Effect, pipe, Record, Array as _Array, Tuple } from "effect";
+import {
+  Effect,
+  pipe,
+  Record,
+  Array as _Array,
+  BigInt as _BigInt,
+  Tuple,
+  Option,
+} from "effect";
 import { Address, Assets, UTxO, Wallet } from "@lucid-evolution/core-types";
 import {
   ERROR_MESSAGE,
   RunTimeError,
-  TransactionError,
   TxBuilderError,
   TxBuilderErrorCause,
 } from "../../Errors.js";
@@ -13,9 +20,7 @@ import * as TxBuilder from "../TxBuilder.js";
 import * as TxSignBuilder from "../../tx-sign-builder/TxSignBuilder.js";
 import {
   assetsToValue,
-  coresToTxOutputs,
   coreToTxOutput,
-  coreToUtxo,
   isEqualUTxO,
   selectUTxOs,
   sortUTxOs,
@@ -45,14 +50,14 @@ const getWalletInfo = (
   config: TxBuilder.TxBuilderConfig,
 ): Effect.Effect<WalletInfo, TxBuilderError, never> =>
   Effect.gen(function* () {
-    const wallet = yield* pipe(
+    const wallet: Wallet = yield* pipe(
       Effect.fromNullable(config.lucidConfig.wallet),
       Effect.orElseFail(() =>
         completeTxError("MissingWallet", ERROR_MESSAGE.MISSING_WALLET),
       ),
     );
-    const address = yield* Effect.promise(() => wallet.address());
-    const inputs = yield* pipe(
+    const address: Address = yield* Effect.promise(() => wallet.address());
+    const inputs: UTxO[] = yield* pipe(
       Effect.tryPromise({
         try: () => wallet.getUtxos(),
         catch: (error) => completeTxError("Provider", String(error)),
@@ -79,7 +84,10 @@ export const complete = (
 
     // Set collateral input if there are script executions
     if (config.scripts.size > 0) {
-      const collateralInput = yield* findCollateral(walletInfo.inputs);
+      const collateralInput = yield* findCollateral(
+        config.lucidConfig.protocolParameters.coinsPerUtxoByte,
+        walletInfo.inputs,
+      );
       setCollateral(config, collateralInput, walletInfo.address);
     }
 
@@ -103,10 +111,15 @@ export const complete = (
     // this is just a patch, and we should find a better way to do this
     config.consumedInputs = [...config.collectedInputs, ...inputsToAdd];
 
-    const txRedeemerBuilder = config.txBuilder.build_for_evaluation(
-      0,
-      CML.Address.from_bech32(walletInfo.address),
-    );
+    const txRedeemerBuilder = yield* Effect.try({
+      try: () =>
+        config.txBuilder.build_for_evaluation(
+          0,
+          CML.Address.from_bech32(walletInfo.address),
+        ),
+      catch: (error) => completeTxError("BuildEvaluation", String(error)),
+    });
+
     if (
       options.localUPLCEval !== false &&
       txRedeemerBuilder.draft_tx().witness_set().redeemers()
@@ -120,12 +133,16 @@ export const complete = (
       CML.Address.from_bech32(walletInfo.address),
       true,
     );
-    const tx = config.txBuilder
-      .build(
-        CML.ChangeSelectionAlgo.Default,
-        CML.Address.from_bech32(walletInfo.address),
-      )
-      .build_unchecked();
+    const tx = yield* Effect.try({
+      try: () =>
+        config.txBuilder
+          .build(
+            CML.ChangeSelectionAlgo.Default,
+            CML.Address.from_bech32(walletInfo.address),
+          )
+          .build_unchecked(),
+      catch: (error) => completeTxError("Build", String(error)),
+    });
 
     const derivedInputs = deriveInputsFromTransaction(tx);
 
@@ -211,12 +228,8 @@ const setCollateral = (
     config.txBuilder.add_collateral(collateralInput);
   }
   const returnassets = pipe(
-    collateralInputs
-      .map((utxo) => utxo.assets)
-      .reduce((acc, cur) =>
-        Record.union(acc, cur, (self, that) => self + that),
-      ),
-    Record.union({ lovelace: -5_000_000n }, (self, that) => self + that),
+    sumAssetsFromInputs(collateralInputs),
+    Record.union({ lovelace: -5_000_000n }, _BigInt.sum),
   );
 
   const collateralOutputBuilder =
@@ -233,12 +246,29 @@ const setCollateral = (
 };
 
 const findCollateral = (
+  coinsPerUtxoByte: bigint,
   inputs: UTxO[],
 ): Effect.Effect<UTxO[], TxBuilderError, never> =>
   Effect.gen(function* () {
-    const selected = selectUTxOs(sortUTxOs(inputs), {
-      lovelace: 5_000_000n,
-    });
+    // NOTE: While the required collateral is 5 ADA, there may be instances where the UTXOs encountered do not contain enough ADA to be returned to the collateral return address.
+    // For example:
+    // A UTXO with 5.5 ADA will result in an error message such as `BabbageOutputTooSmallUTxO`, since only 0.5 ADA would be returned to the collateral return address.
+    const minLovelaceChangeAddress: Assets = {
+      lovelace: calculateMinLovelace(coinsPerUtxoByte),
+    };
+    const collateralLovelace: Assets = { lovelace: 5_000_000n };
+    const requiredAssets = Record.union(
+      minLovelaceChangeAddress,
+      collateralLovelace,
+      _BigInt.sum,
+    );
+    // const selected = selectUTxOs(sortUTxOs(inputs), requiredAssets);
+    const selected = recursive(
+      sortUTxOs(inputs),
+      collateralLovelace,
+      coinsPerUtxoByte,
+    );
+
     if (_Array.isEmptyArray(selected))
       yield* completeTxError(
         "MissingCollateral",
@@ -247,7 +277,7 @@ const findCollateral = (
     if (selected.length > 3)
       yield* completeTxError(
         "MissingCollateral",
-        `Selected ${selected.length} inputs as collateral, but max collateral inputs is 3 to cover the 5 ADA collateral`,
+        `Selected ${selected.length} inputs as collateral, but max collateral inputs is 3 to cover the 5 ADA collateral ${stringify(selected)}`,
       );
     return selected;
   });
@@ -258,40 +288,40 @@ const coinSelection = (
 ): Effect.Effect<UTxO[], TxBuilderError> =>
   Effect.gen(function* () {
     // NOTE: This is a fee estimation. If the amount is not enough, it may require increasing the fee.
-    const fee: Assets = { lovelace: config.txBuilder.min_fee(false) };
-    const minAdaChangeAddress = {
-      lovelace: calculateMinLovelace(
-        config.lucidConfig.protocolParameters.coinsPerUtxoByte,
-      ),
-    };
-    const negatedMintedAssets = Record.map(
-      config.mintedAssets,
-      (amount) => -amount,
+    // const estimatedFee: Assets = { lovelace: config.txBuilder.min_fee(false) };
+    // TODO: add fee multiplier
+    const estimatedFee: Assets = { lovelace: 2_000_000n };
+    const negatedMintedAssets = negateAssets(config.mintedAssets);
+    const negatedCollectedAssets = negateAssets(
+      sumAssetsFromInputs(config.collectedInputs),
     );
-    const collectedAssets = _Array.isEmptyArray(config.collectedInputs)
-      ? {}
-      : config.collectedInputs
-          .map((utxo) => utxo.assets)
-          .reduce((acc, cur) =>
-            Record.union(acc, cur, (self, that) => self + that),
-          );
-    const negatedCollectedAssets = Record.map(
-      collectedAssets,
-      (amount) => -amount,
-    );
-    const requiredAssets = pipe(
+    // Calculate the net change in assets (delta)
+    const assetsDelta: Assets = pipe(
       config.totalOutputAssets,
-      Record.union(minAdaChangeAddress, (self, that) => self + that),
-      Record.union(fee, (self, that) => self + that),
-      Record.union(negatedCollectedAssets, (self, that) => self + that),
-      Record.union(negatedMintedAssets, (self, that) => self + that),
+      Record.union(estimatedFee, _BigInt.sum),
+      Record.union(negatedCollectedAssets, _BigInt.sum),
+      Record.union(negatedMintedAssets, _BigInt.sum),
+    );
+    // Filter and obtain only the required assets (those with a positive amount)
+    const requiredAssets = pipe(
+      assetsDelta,
       Record.filter((amount) => amount > 0n),
     );
-
     // No UTxOs need to be selected if collected inputs are sufficient
     if (Record.isEmptyRecord(requiredAssets)) return [];
-
-    const selected = selectUTxOs(sortUTxOs(availableInputs), requiredAssets);
+    // Filter and obtain assets that are present in the inputs and mints but are not required by the outputs
+    // Negate these assets to get their positive amounts
+    const notRequiredAssets = pipe(
+      assetsDelta,
+      Record.filter((amount) => amount < 0n),
+      negateAssets,
+    );
+    const selected = recursive(
+      sortUTxOs(availableInputs),
+      requiredAssets,
+      config.lucidConfig.protocolParameters.coinsPerUtxoByte,
+      notRequiredAssets,
+    );
     if (_Array.isEmptyArray(selected))
       yield* completeTxError(
         "NotFound",
@@ -308,7 +338,6 @@ const evalTransaction = (
   Effect.gen(function* () {
     //FIX: this returns undefined
     const txEvaluation = setRedeemerstoZero(txRedeemerBuilder.draft_tx())!;
-    // console.log(txEvaluation?.to_json());
     const txUtxos = [
       ...walletInputs,
       ...config.collectedInputs,
@@ -339,6 +368,7 @@ const evalTransaction = (
 
 const calculateMinLovelace = (
   coinsPerUtxoByte: bigint,
+  multiAssets?: Assets,
   changeAddress?: string,
 ) => {
   const dummyAddress =
@@ -348,7 +378,12 @@ const calculateMinLovelace = (
       CML.Address.from_bech32(changeAddress ? changeAddress : dummyAddress),
     )
     .next()
-    .with_asset_and_min_required_coin(CML.MultiAsset.new(), coinsPerUtxoByte)
+    .with_asset_and_min_required_coin(
+      multiAssets
+        ? assetsToValue(multiAssets).multi_asset()
+        : CML.MultiAsset.new(),
+      coinsPerUtxoByte,
+    )
     .build()
     .output()
     .amount()
@@ -369,4 +404,85 @@ const deriveInputsFromTransaction = (tx: CML.Transaction) => {
     utxos.push(utxo);
   }
   return utxos;
+};
+
+/**
+ * Returns a new `Assets`
+ *
+ * Negates the amounts of all assets in the given record.
+ */
+const negateAssets = (assets: Assets): Assets =>
+  Record.map(assets, (amount) => -amount);
+
+/**
+ * Returns a new Assets
+ *
+ * Sums the assets from an array of UTxO inputs.
+ */
+const sumAssetsFromInputs = (inputs: UTxO[]) =>
+  _Array.isEmptyArray(inputs)
+    ? {}
+    : inputs
+        .map((utxo) => utxo.assets)
+        .reduce((acc, cur) => Record.union(acc, cur, _BigInt.sum));
+
+const calculateExtraLovelace = (
+  leftoverAssets: Assets,
+  coinsPerUtxoByte: bigint,
+): Option.Option<Assets> => {
+  return pipe(leftoverAssets, (assets) => {
+    const minLovelace = calculateMinLovelace(coinsPerUtxoByte, assets);
+    const currentLovelace = assets["lovelace"] || 0n;
+    return currentLovelace > minLovelace
+      ? Option.none()
+      : Option.some({ lovelace: minLovelace - currentLovelace });
+  });
+};
+
+export const recursive = (
+  inputs: UTxO[],
+  requiredAssets: Assets,
+  coinsPerUtxoByte: bigint,
+  externalAssets: Assets = {},
+): UTxO[] => {
+  let selected = selectUTxOs(inputs, requiredAssets);
+  if (_Array.isEmptyArray(selected)) return [];
+  const selectedAssets: Assets = sumAssetsFromInputs(selected);
+  let availableAssets: Assets = pipe(
+    selectedAssets,
+    Record.union(requiredAssets, (self, that) => self - that),
+    Record.union(externalAssets, _BigInt.sum),
+  );
+
+  let extraLovelace: Assets | undefined = pipe(
+    calculateExtraLovelace(availableAssets, coinsPerUtxoByte),
+    Option.getOrUndefined,
+  );
+  let remainingInputs: UTxO[] = _Array.differenceWith(isEqualUTxO)(
+    inputs,
+    selected,
+  );
+  while (extraLovelace) {
+    const extraSelected = selectUTxOs(remainingInputs, extraLovelace);
+    if (_Array.isEmptyArray(extraSelected)) {
+      return [];
+    }
+    const extraSelectedAssets: Assets = sumAssetsFromInputs(extraSelected);
+    selected = [...selected, ...extraSelected];
+    availableAssets = Record.union(
+      availableAssets,
+      extraSelectedAssets,
+      _BigInt.sum,
+    );
+
+    extraLovelace = pipe(
+      calculateExtraLovelace(availableAssets, coinsPerUtxoByte),
+      Option.getOrUndefined,
+    );
+    remainingInputs = _Array.differenceWith(isEqualUTxO)(
+      remainingInputs,
+      selected,
+    );
+  }
+  return selected;
 };